// frontend/src/app/api/models/session.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { Permissions } from '../models/permissions';

/**
 * The role the caller currently acts as, plus its permissions.
 */
export interface Session {
  isAnonymous: boolean;
  permissions: Permissions;
  roleId: string;
  roleName: string;
}